+++ components/_util/reactNode.ts
@@ -13,7 +13,10 @@ export function replaceElement(
 ): React.ReactNode {
   if (!isValidElement(element)) return replacement;
 
-  return React.cloneElement(element, typeof props === 'function' ? props(element.props) : props);
+  return React.cloneElement(
+    element,
+    typeof props === 'function' ? props(element.props || {}) : props,
+  );
 }
 
 export function cloneElement(element: React.ReactNode, props?: RenderProps): React.ReactElement {
