+++ components/style/themes/default.less
@@ -860,7 +860,7 @@
 
 // Skeleton
 // ---
-@skeleton-color: rgba(190, 190, 190, .2);
+@skeleton-color: rgba(190, 190, 190, 0.2);
 @skeleton-to-color: shade(@skeleton-color, 5%);
 @skeleton-paragraph-margin-top: 28px;
 @skeleton-paragraph-li-margin-top: @margin-md;
