+++ package.json
@@ -129,7 +129,7 @@
     "rc-input-number": "~6.1.0",
     "rc-mentions": "~1.5.0",
     "rc-menu": "~8.10.0",
-    "rc-motion": "^2.2.0",
+    "rc-motion": "^2.4.0",
     "rc-notification": "~4.5.2",
     "rc-pagination": "~3.1.0",
     "rc-picker": "~2.4.1",
@@ -235,7 +235,7 @@
     "open": "^7.0.3",
     "preact": "^10.0.0",
     "preact-compat": "^3.18.5",
-    "prettier": "^2.0.1",
+    "prettier": "^2.2.0",
     "pretty-quick": "^3.0.0",
     "querystring": "^0.2.0",
     "rc-footer": "^0.6.3",
@@ -271,7 +271,7 @@
     "reqwest": "^2.0.5",
     "rimraf": "^3.0.0",
     "scrollama": "^2.0.0",
-    "simple-git": "^2.0.0",
+    "simple-git": "^2.23.0",
     "string-replace-loader": "^2.3.0",
     "stylelint": "^13.0.0",
     "stylelint-config-prettier": "^8.0.0",
